Delete extern

+++ common/utils.c
@@ -1,6 +1,6 @@
 #include <common.h>
 
-void itostring(uint8_t digit, uint8_t data, uint8_t* buf)
+void itostring(uint8_t digit, uint16_t data, uint8_t* buf)
 {
     uint8_t lc;
